Extracted parser's patch to the separate file.
