+++ lib/opal/parser/patch.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+if RUBY_ENGINE == 'opal'
+  class Parser::Lexer::Literal
+    undef :extend_string
+
+    def extend_string(string, ts, te)
+      @buffer_s ||= ts
+      @buffer_e = te
+
+      # Patch for opal-parser, original:
+      # @buffer << string
+      @buffer += string
+    end
+  end
+end
